+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Support for ESLint 10
+- Support for TypeScript 6
+
+### Fixed
+- **BREAKING**: commonjs export is now using default export.
+```diff
+-const { default: dczajkowski } = require('@dczajkowski/eslint-plugin')
++const dczajkowski = require('@dczajkowski/eslint-plugin')
+````
+
 ## [1.1.0] - 2026-01-09
 
 ### Added